Update qtrServices etc.
