+++ modules/custom/qtrUser/include/form_alter.inc
@@ -8,51 +8,8 @@
  * Implements hook_form_FORM_ID_alter() for user_profile_form.
  */
 function qtrUser_form_user_profile_form_alter(&$form, &$form_state, $form_id) {
-  // Replace 'field_projects' with a list of subscribed projects.
-  $arr_projects = field_get_items('user', $form['#user'], 'field_projects');
-  if ($arr_projects) {
-    $lng = $form['#user']->translation_lng;
-    $base_url = qtr::utils_get_client_url($lng);
-    $items = [];
-    foreach ($arr_projects as $proj) {
-      $project = $proj['value'];
-      $url = "$base_url/qtr/project/$project/$lng/dashboard";
-      $items[] = "<a href='$url' class='btn btn-default' target='_blank'>$project/$lng</a>";
-    }
-    $form['field_projects'] = [
-      '#type'  => 'fieldset',
-      '#title' => t('Subscribed Projects'),
-      'subscribed_projects' => [
-        '#markup' => implode(' ', $items),
-      ],
-    ];
-  }
-  else {
-    $form['field_projects']['#access'] = user_access('qtranslate-admin');
-  }
-
   // Hide 'field_auxiliary_languages', it is not useful yet.
   $form['field_auxiliary_languages']['#access'] = FALSE;
-
-  // Add fields for admin_projects and moderate_projects.
-  $form['admin'] = array(
-    '#type'  => 'fieldset',
-    '#title' => t('Fields that can be seen and modified only by an administrator'),
-    '#access' => in_array('administrator', $GLOBALS['user']->roles),
-    '#weight' => -100,   // display it on the top
-  );
-  $form['admin']['admin_projects'] = array(
-    '#type' => 'textarea',
-    '#title' => t('The list of projects that user can administrate'),
-    '#description' => t("Set the projects that the user can admin. Enter each project on a separate line, in the form <em>origin/project</em>."),
-    '#default_value' => implode("\n", $form_state['user']->admin_projects),
-  );
-  $form['admin']['moderate_projects'] = array(
-    '#type' => 'textarea',
-    '#title' => t('The list of projects that user can moderate'),
-    '#description' => t("Set the projects that the user can moderate. Enter each project on a separate line, in the form <em>origin/project</em>."),
-    '#default_value' => implode("\n", $form_state['user']->moderate_projects),
-  );
 }
 
 /**
